src/requirements.txt
django-crispy-forms
requests